% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.R
\name{deg_2_km}
\alias{deg_2_km}
\title{Distances in degree to kilometer}
\usage{
deg_2_km(x, start, end)
}
\arguments{
\item{x}{numeric vector of distances in degree}

\item{start, end}{start and end point as vectors with lon, lat}
}
\value{
numeric vector
}
\description{
Converts distances along a great circle path from degree into kilometer
}
