% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_points.R
\name{draw}
\alias{draw}
\alias{get_coordinates}
\alias{draw_profile}
\title{Draw a profile line or a point to retrieve coordinates}
\usage{
get_coordinates(x, n = 1, type = "o", col = "#B63679FF", ...)

draw_profile(x, n = 10, ...)
}
\arguments{
\item{x}{\code{sf} object}

\item{n}{the maximum number of points to locate.  Valid values start
    at 1.}

\item{type}{One of \code{"n"}, \code{"p"}, \code{"l"} or
    \code{"o"}. If \code{"p"} or \code{"o"} the points are
    plotted; if \code{"l"}  or \code{"o"} they are joined by lines.}

\item{col}{color of line or point}

\item{...}{additional graphics parameters used if \code{type != "n"}
    for plotting the locations.}
}
\description{
Draw a profile line or a point to retrieve coordinates
}
