% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprofiler-package.R
\docType{package}
\name{geoprofiler-package}
\alias{geoprofiler-package}
\title{library(geoprofiler)}
\description{
Measures accurate spatial distances along a profile that is specified
either by two locations or by a orientation. Also provides functions to
produce swath profiles.
}
\note{
A list of documented functions may be viewed by typing
\code{help(package="geoprofiler")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/geoprofiler/}
  \item Report bugs at \url{https://github.com/tobiste/geoprofiler/issues}
}

}
\author{
Tobias Stephan
}
\keyword{internal}
