% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_points.R
\name{line_ends}
\alias{line_ends}
\title{#' Distances in degree to kilometer
#'
#' Converts distances along a great circle path from degree into kilometer
#'
#' @param x numeric vector of distances in degree
#' @param start,end start and end point as vectors with lon, lat
#' @return units object
#' @importFrom scales rescale
#' @importFrom units set_units
#' @export
deg_2_km <- function(x, start, end) {
distance.km.total <- point_distance(start, end)
distances.km <- scales::rescale(x, to = c(0, distance.km.total))
units::set_units(distances.km, "km")
}
Extract end points of a line}
\usage{
line_ends(x)
}
\arguments{
\item{x}{\code{sf} line object}
}
\value{
\code{sf} point object
}
\description{
#' Distances in degree to kilometer
#'
#' Converts distances along a great circle path from degree into kilometer
#'
#' @param x numeric vector of distances in degree
#' @param start,end start and end point as vectors with lon, lat
#' @return units object
#' @importFrom scales rescale
#' @importFrom units set_units
#' @export
deg_2_km <- function(x, start, end) {
distance.km.total <- point_distance(start, end)
distances.km <- scales::rescale(x, to = c(0, distance.km.total))
units::set_units(distances.km, "km")
}
Extract end points of a line
}
\examples{
p1 <- data.frame(lon = -90.8, lat = 48.6) |>
  sf::st_as_sf(coords = c("lon", "lat"), crs = "WGS84")
profile_points(p1,
  profile.azimuth = 135, profile.length = 10000,
  crs = sf::st_crs("EPSG:26915")
) |>
  profile_line() |>
  line_ends()
}
