% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations_example}
\alias{locations_example}
\title{Example \code{sf} data set}
\format{
An object of class \code{sf}
}
\usage{
data('locations_example')
}
\description{
example dataset
}
\examples{
data("locations_example")
head(locations_example)
}
\keyword{datasets}
