% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_example}
\alias{raster_example}
\title{Example raster data set}
\format{
An object of class \code{matrix}
}
\usage{
data('raster_example')
}
\description{
example dataset
}
\examples{
data("raster_example")
head(raster_example)
}
\keyword{datasets}
