% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.R
\name{swathR}
\alias{swathR}
\title{Swath Profile}
\usage{
swathR(coords, raster, k, dist, crs, method)
}
\arguments{
\item{coords}{matrix(ncol=2, nrow=2) with x and y coordinates of beginning and
end point of the baseline; each point in one row
\describe{
\item{column 1}{xcoordinates}
\item{column 2}{ycoordinates}
}}

\item{raster}{raster file (loaded with \code{\link[raster:raster]{raster::raster()}})}

\item{k}{integer; number of lines on each side of the baseline}

\item{dist}{numeric; distance between lines}

\item{crs}{string; CRS}

\item{method}{string; method for extraction of raw data, see
\code{\link[raster:extract]{raster::extract()}}: default value: "bilinear"}
}
\description{
Calculate swath-profile values perpendicular to a straight baseline. The
baseline is generated between two user-defined points (X|Y), see argument
\code{coords}. The distance between samples and the number of samples can be
specified, see arguments \code{k} and \code{dist}. Values of the swath-profile are
extracted from a given raster file, see argument \code{raster}. CRS of raster
and points have to be the same.
}
\author{
V. Haburaj
}
