% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{vcross}
\alias{vcross}
\title{Vector cross-product}
\usage{
vcross(x, y)
}
\arguments{
\item{x, y}{numeric 3-column matrices representing vectors}
}
\value{
numeric 3-column matrix
}
\description{
Vector cross-product
}
\keyword{internal}
