% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{vec2mat}
\alias{vec2mat}
\title{Helper function to format vectors into 3-column matrices}
\usage{
vec2mat(x)
}
\arguments{
\item{x}{numeric 3-column matrix or 3-element vector}
}
\value{
numeric 3-column matrix
}
\description{
Helper function to format vectors into 3-column matrices
}
\keyword{internal}
