% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{vnorm}
\alias{vnorm}
\title{Vector normalisation}
\usage{
vnorm(x)
}
\arguments{
\item{x}{numeric 3-column matrix}
}
\value{
numeric 3-column matrix
}
\description{
Vector normalisation
}
\keyword{internal}
