% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{vrotate}
\alias{vrotate}
\title{Vector rotation}
\usage{
vrotate(x, rotaxis, rotangle)
}
\arguments{
\item{x, rotaxis}{numeric 3-column matrices representing vectors}

\item{rotangle}{numeric. Angle in radians}
}
\value{
numeric 3-column matrix
}
\description{
Vector rotation
}
\keyword{internal}
