% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprofiler-package.R
\docType{package}
\name{geoprofiler-package}
\alias{geoprofiler}
\alias{geoprofiler-package}
\title{geoprofiler: Distance Along Profile}
\description{
Toolset to measures distances along lines defined by a direction or create swath profiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/geoprofiler/}
  \item Report bugs at \url{https://github.com/tobiste/geoprofiler/issues}
}

}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

}
\keyword{internal}
