% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_points.R
\name{profile_azimuth}
\alias{profile_azimuth}
\title{Azimuth Between Profile Points}
\usage{
profile_azimuth(profile)
}
\arguments{
\item{profile}{\code{sf} point object. First point marks the start point.}
}
\value{
numeric. Azimuth in degrees
}
\description{
Azimuth Between Profile Points
}
\examples{
p1 <- data.frame(lon = -90.8, lat = 48.6) |>
  sf::st_as_sf(coords = c("lon", "lat"), crs = "WGS84")

profile_points(p1, profile.azimuth = 135, profile.length = 10000,
  crs = sf::st_crs("EPSG:26915")) |>
  profile_azimuth()
}
