% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.R
\name{swath_profile}
\alias{swath_profile}
\title{Elevation profile}
\usage{
swath_profile(x)
}
\arguments{
\item{x}{list. an return object from \code{swathR}}
}
\value{
tibble
}
\description{
Extracts the minimum and maximum elevation data along a swathR profile.
}
